refactor(cli): Move WSDL parser utility to helpers module
