+++ onvif/cli/helpers/__init__.py
@@ -4,6 +4,10 @@ from onvif.cli.helpers.discovery import discover_devices, select_device_interact
 from onvif.cli.helpers.execution import execute_command, setup_warning_format
 from onvif.cli.helpers.files import save_output_to_file
 from onvif.cli.helpers.products import search_products
+from onvif.cli.helpers.wsdl_parser import (
+    get_method_documentation,
+    get_operation_type_info,
+)
 
 __all__ = [
     "search_products",
@@ -12,4 +16,6 @@ __all__ = [
     "select_device_interactive",
     "execute_command",
     "setup_warning_format",
+    "get_method_documentation",
+    "get_operation_type_info",
 ]
